Fixup mm headers
Commit 822fd27548c495 defined that `mm.h` and `dymelor.h` should be the
external and internal mm headers.

This commit re-enforces this.
