+++ src/mm/dymelor.h
@@ -140,6 +140,43 @@ typedef struct _malloc_state malloc_state;
 #define get_area_top_pointer(ptr) ( (malloc_area **)(*get_top_pointer(ptr)) )
 #define get_area(ptr) ( *(get_area_top_pointer(ptr)) )
 
+#define PER_LP_PREALLOCATED_MEMORY (262144L * PAGE_SIZE)	// This should be power of 2 multiplied by a page size. This is 1GB per LP.
+#define BUDDY_GRANULARITY PAGE_SIZE	// This is the smallest chunk released by the buddy in bytes. PER_LP_PREALLOCATED_MEMORY/BUDDY_GRANULARITY must be integer and a power of 2
+
+
+struct segment {
+	unsigned char *base;
+	unsigned char *brk;
+};
+
+extern size_t __page_size;
+#define PAGE_SIZE ({ \
+			if(unlikely(__page_size == 0))\
+				__page_size = getpagesize();\
+			__page_size;\
+		  })
+
+struct slab_header {
+#ifndef NDEBUG
+	atomic_t presence;
+#endif
+	struct slab_header *prev, *next;
+	uint64_t slots;
+	uintptr_t refcount;
+	struct slab_header *page;
+	uint8_t data[] __attribute__((aligned(sizeof(void *))));
+};
+
+struct slab_chain {
+	spinlock_t lock;
+	size_t itemsize, itemcount;
+	size_t slabsize, pages_per_alloc;
+	uint64_t initial_slotmask, empty_slotmask;
+	uintptr_t alignment_mask;
+	struct slab_header *partial, *empty, *full;
+};
+
+
 /***************
  * EXPOSED API *
  ***************/
@@ -158,20 +195,32 @@ extern void do_free(struct lp_struct *, void *ptr);
 extern void *allocate_lp_memory(struct lp_struct *, size_t);
 extern void free_lp_memory(struct lp_struct *, void *);
 
-// Checkpointing API
-extern void *log_full(struct lp_struct *);
-extern void *log_state(struct lp_struct *);
-extern void log_restore(struct lp_struct *, state_t *);
-extern void log_delete(void *);
 
 // Userspace API
 extern void *__wrap_malloc(size_t);
 extern void __wrap_free(void *);
 extern void *__wrap_realloc(void *, size_t);
 extern void *__wrap_calloc(size_t, size_t);
-extern void clean_buffers_on_gvt(struct lp_struct *, simtime_t);
 
-extern void ecs_init(void);
+
+/***************************
+ * BUDDY SYSTEM
+ ***************************/
+
+// buddy block size expressed in 2^n, e.g.: BUDDY_BLOCK_SIZE_EXP = 4, block_size = 16 TODO make this the pagesize
+#define BUDDY_BLOCK_SIZE_EXP 12
+
+struct buddy {
+	spinlock_t lock;
+	size_t size;
+	size_t longest[] __attribute__((aligned(sizeof(size_t)))); // an array based binary tree
+};
+
+extern struct buddy *buddy_new(size_t requested_size);
+extern void buddy_destroy(struct buddy *self);
+extern void *allocate_buddy_memory(struct buddy *self, void *base_mem, size_t requested_size);
+extern void free_buddy_memory(struct buddy *self, void *base_mem, void *ptr);
+
 
 // This is used to help ensure that the platform is not using malloc.
 #pragma GCC poison malloc free realloc calloc
