+++ src/mm/mm.h
@@ -33,50 +33,17 @@
 #include <sys/mman.h>
 #include <pthread.h>
 #include <core/core.h>
+#include <scheduler/process.h>
 #include <arch/atomic.h>
-#include <mm/dymelor.h>
 
-struct segment {
-	unsigned char *base;
-	unsigned char *brk;
-};
-
-extern size_t __page_size;
-#define PAGE_SIZE ({ \
-			if(unlikely(__page_size == 0))\
-				__page_size = getpagesize();\
-			__page_size;\
-		  })
-
-struct slab_header {
-#ifndef NDEBUG
-	atomic_t presence;
-#endif
-	struct slab_header *prev, *next;
-	uint64_t slots;
-	uintptr_t refcount;
-	struct slab_header *page;
-	uint8_t data[] __attribute__((aligned(sizeof(void *))));
-};
-
-struct slab_chain {
-	spinlock_t lock;
-	size_t itemsize, itemcount;
-	size_t slabsize, pages_per_alloc;
-	uint64_t initial_slotmask, empty_slotmask;
-	uintptr_t alignment_mask;
-	struct slab_header *partial, *empty, *full;
-};
+struct _malloc_state;
 
 struct memory_map {
-	malloc_state *m_state;
+	struct _malloc_state *m_state;
 	struct buddy *buddy;
 	struct slab_chain *slab;
 	struct segment *segment;
 };
-
-#define PER_LP_PREALLOCATED_MEMORY (262144L * PAGE_SIZE)	// This should be power of 2 multiplied by a page size. This is 1GB per LP.
-#define BUDDY_GRANULARITY PAGE_SIZE	// This is the smallest chunk released by the buddy in bytes. PER_LP_PREALLOCATED_MEMORY/BUDDY_GRANULARITY must be integer and a power of 2
 
 extern bool allocator_init(void);
 extern void allocator_fini(void);
@@ -90,6 +57,14 @@ extern void finalize_memory_map(struct lp_struct *lp);
 extern struct slab_chain *slab_init(const size_t itemsize);
 extern void *slab_alloc(struct slab_chain *);
 extern void slab_free(struct slab_chain *, const void *const addr);
+
+
+// Checkpointing API
+extern void *log_full(struct lp_struct *);
+extern void *log_state(struct lp_struct *);
+extern void log_restore(struct lp_struct *, state_t *);
+extern void log_delete(void *);
+extern void clean_buffers_on_gvt(struct lp_struct *, simtime_t);
 
 /* Simulation Platform Memory APIs */
 extern inline void *rsalloc(size_t);
